Finished text prompt on item pickup, and cleaned up functionality of organizer puzzle.
When you pick up an item and its placed into your inventory, a text will appear on screen indicating to the player that they have collected a particular item/artifact. I also made the highlight interactable item on hover a bit more transparent so its not overdoing it. And finally, for the organizer puzzle, only when all 3 items are placed in the correct sequence, they are locked into place.