Added remaining trinket boxes, fixed jumpscare quality, etc.
Added the 2nd and 3rd trinket boxes that will hold the keys to unlock each section of the maze. Fixed the issue with the jumpscare quality for higher scalability settings, it is now high quality across all settings. Configured the decal size for the 2nd and 3rd keys. Added the 2nd and 3rd trinket boxes into map for testing.